refactor: eulexia fab actions divided onto components. using a context for the ruler feature
